Update mm1
Hi mr. van Foreest,

I thought this might be better since L(t) has to be non-negative, not necessarily positive. Of course, for someone to depart L(t) must be positive, but for the departure process to be given by a poisson process, L(t) does not have to be positive all the time. I understand now what you mean, but it was quite confusing to read this. You could probably also rewrite the whole sentence, but this might be an easy fix.

Best,
Thierry Volker, s2941643