Update pasta
Hi mr van Foreest,

The word arbitrary here is wrong in my humble opinion. It makes it feel like the people that are already in the system can also observe n people in the system at the moment someone else arrives, which would also count for pi(n). I think that what you mean is that the person that arrives sees n people, and not that any person sees n people in the system as an arbitrary person arrives. It is a subtle difference, but I think it is more clear this way. 

Best,

Thierry Volker, s2941643