Update solution ex. 13
Hi mr. van Foreest,

I changed the 't' to 's', since the whole text the time variable is 's' and there is no reason, I think, that this should be a t. It was a bit confusing for me since I thought you where referring to something else.

Best,
Thierry Volker, s2941643