Update solution level crossing
Hi mr. van Foreest,

I think this is correct. In the old description it effectively said that the first arrival time A1 is 2, but in the question it says that the first arrival is at 0. So you could leave the arrival at time 0 out or do this, in my opinion.

Best,
Thierry Volker, s2941643