Logical change up-crossing condition
I think that this is logically necessary. If Dk-1 leaves an empty system behind, but Yk <n, then L(Dk) < n. This mean that both sides are 0 and hence equal to each other, which would imply an up-crossing of level n. But this is false of course. Therefore I added the addition.

Best,
Thierry Volker, s2941643