Sentence change
Hi mr. van Foreest,

I adjusted this since the question is in my opinion about the average time a job in a batch has to wait until the batch it is in is completed. In the solution the difference is well contrasted. B-1/lambda is the total time to form a batch, and it is divided by two in order to get the average waiting time of a job.

Best,
Thierry Volker, s2941643